[Server] Rename the server lua api files to server_lua_api.h from lua_api.h
