+++ src/server/lua/data_api.cpp
@@ -1,4 +1,4 @@
-#include "lua_api.h"
+#include "server_lua_api.h"
 
 #include <common/scripting/script_engine.h>
 #include <common/world/biome.h>
